handle ISOs that have a cdrom.source

+++ subiquity/server/apt.py
@@ -67,13 +67,13 @@ def get_index_targets() -> List[str]:
     return [key for key in targets if key.count("::") == 3]
 
 
-def apt_sourceparts_files(mp: Mountpoint) -> List[str]:
+def apt_sourceparts_files(mp: Mountpoint) -> List[pathlib.Path]:
     # Return the relative path of the files from
     # mp("etc/apt/sources.list.d") that apt will read.
     root = mp.pp()
     sources_list_d = root / "etc/apt/sources.list.d"
     paths = list(sources_list_d.glob("*.list")) + list(sources_list_d.glob("*.sources"))
-    return [str(p.relative_to(root)) for p in paths]
+    return [p.relative_to(root) for p in paths]
 
 
 class AptConfigurer:
@@ -253,15 +253,28 @@ class AptConfigurer:
                 missing_ok=True
             )
             for relpath in apt_sourceparts_files(self.configured_tree):
+                # If there is an apt source called cdrom.sources since it's
+                # the file we use, we assume this is an ISO for 26.04 which
+                # has the ISO pool enabled in the squashfs and so leave it in
+                # place.
+                if relpath.name == "cdrom.sources":
+                    continue
                 self.install_tree.pp(relpath).unlink()
 
         codename = lsb_release(dry_run=self.app.opts.dry_run)["codename"]
 
-        write_file(
-            self.install_tree.p("etc/apt/sources.list"),
-            f"deb [check-date=no] file:///cdrom {codename} main restricted\n",
-            mode=0o644,
-        )
+        new_style_iso = self.configured_tree.pp(
+            "etc/apt/sources.list.d/cdrom.sources"
+        ).exists()
+
+        if not new_style_iso:
+            # If we _didn't_ find a cdrom.sources, we need to add the ISO pool
+            # as a package source.
+            write_file(
+                self.install_tree.p("etc/apt/sources.list"),
+                f"deb [check-date=no] file:///cdrom {codename} main restricted\n",
+                mode=0o644,
+            )
 
         # workaround LP: #2105480 and many many many like it
         apt_lists = self.install_tree.pp("var/lib/apt/lists")
@@ -326,7 +339,7 @@ class AptConfigurer:
                 ]
             )
 
-        def _restore_file(path: str) -> None:
+        def _restore_file(path: str | pathlib.Path) -> None:
             shutil.copyfile(self.configured_tree.p(path), target_mnt.p(path))
 
         # The file only exists if we are online
@@ -355,6 +368,9 @@ class AptConfigurer:
             for relpath in apt_sourceparts_files(self.configured_tree):
                 _restore_file(relpath)
             await _restore_dir("var/lib/apt/lists")
+
+        # If the ISO came with a cdrom.sources, delete it.
+        target_mnt.pp("etc/apt/sources.list.d/cdrom.sources").unlink(missing_ok=True)
 
         await self.cleanup()
         try:
